refactor: JSON을 반환하는 체스 방 관련된 요청을 @RestController로 리팩토링 및 클래스 분리
